refactor: 사다리 생성 로직 LadderGame으로 이동
